+++ src/main/java/com/icebox/freshmate/domain/grocery/application/GroceryExpirationStrategy.java
@@ -1,0 +1,8 @@
+package com.icebox.freshmate.domain.grocery.application;
+
+import java.time.LocalDate;
+
+public interface GroceryExpirationStrategy {
+
+	void handleGrocery(LocalDate currentDate);
+}
